feat!: make current thread-count accessible in slice-workers. (#851)
Threads started for working on an entry in a slice can now see the amount
of threads left for use (and manipulate that variable) which effectively
allows them to implement their own parallelization on top of the current one.

This is useful if there is there is very imbalanced work within the slice itself.

While at it, we also make consumer functions mutable as they exsit per thread.