Propogate value errors to user
This should provide more meaningful errors to the user and is somewhat
closer to how I would expect a parser to work in a situation like this
(ie we should ensure `cut` errors get propagated).

The main risk is if there is any error recovery meant to be going on.