chore!: restructure gix-url parsing error variants
All error variants now own a copy of the input which is used in their
display implementation. I noticed that URL parsing errors are almost never
acted upon but only wrapped and bubbled up to the user. It therefore
makes sense to ensure the message for this error is informative enough
by default.

If an error can be thrown for different types of URLs it now also
includes a kind field. With this field the caller can determine what
kind of URL we tried to parse. Additionally, this information is used
for the error message too.

For testing the assert_matches crate was added which implements the
unstable feature with the same name in stable rust. It makes testing the
invalid variants much more convenient.

BREAKING because public fields of some error variants are now no longer
available